feat: update home to show two tabs, update detail styles
