+++ package.json
@@ -23,21 +23,23 @@
     "@supabase/supabase-js": "^2.100.0",
     "autoskills": "^0.2.7",
     "babel-preset-expo": "~55.0.8",
-    "expo": "~55.0.8",
+    "expo": "~55.0.15",
     "expo-blur": "~55.0.14",
-    "expo-font": "~55.0.4",
-    "expo-linear-gradient": "~55.0.11",
-    "expo-localization": "~55.0.9",
-    "expo-status-bar": "~55.0.4",
+    "expo-font": "~55.0.6",
+    "expo-linear-gradient": "~55.0.13",
+    "expo-localization": "~55.0.13",
+    "expo-status-bar": "~55.0.5",
     "i18next": "^25.10.9",
     "phosphor-react-native": "^3.0.3",
     "react": "19.2.0",
     "react-i18next": "^16.6.6",
-    "react-native": "0.83.2",
+    "react-native": "0.83.4",
+    "react-native-pager-view": "8.0.0",
     "react-native-paper": "^5.15.0",
     "react-native-safe-area-context": "~5.6.2",
     "react-native-screens": "~4.23.0",
     "react-native-svg": "15.15.3",
+    "react-native-tab-view": "^4.3.0",
     "react-native-url-polyfill": "^3.0.0",
     "styled-components": "^6.3.12"
   },
