+++ src/components/MoviesList/types.ts
@@ -1,0 +1,6 @@
+import type { UserSeries } from '@/types/database.types';
+
+export type MoviesListProps = {
+  userMovies: UserSeries[];
+  isLoading: boolean;
+};
