+++ src/components/SeriesList/types.ts
@@ -1,0 +1,6 @@
+import type { UserSeries } from '@/types/database.types';
+
+export type SeriesListProps = {
+  userSeries: UserSeries[];
+  isLoading: boolean;
+};
