+++ src/views/home/viewmodel.ts
@@ -1,11 +1,6 @@
 import { useAuth } from '@/context/AuthContext';
-import { useSeries } from '@/context/SeriesContext';
 import { supabase } from '@/lib/supabase';
-import { ScreenType, TabParamsList } from '@/navigation/types';
-import { SeriesStatus } from '@/types/database.types';
-import { BottomTabNavigationProp } from '@react-navigation/bottom-tabs';
-import { useNavigation } from '@react-navigation/native';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 const getGreetingKey = (): 'morning' | 'afternoon' | 'evening' => {
   const hour = new Date().getHours();
@@ -15,25 +10,22 @@ const getGreetingKey = (): 'morning' | 'afternoon' | 'evening' => {
 };
 
 export const useViewModel = () => {
-  const { session } = useAuth();
-  const { userSeries } = useSeries();
-  const [activeStatus, setActiveStatus] = useState<SeriesStatus>(
-    SeriesStatus.Watching,
-  );
-
-  const [isLoading, setIsLoading] = useState(true);
+  const { session, userName } = useAuth();
   const [firstName, setFirstName] = useState<string | null>(null);
-  const navigation = useNavigation<BottomTabNavigationProp<TabParamsList>>();
 
-  const handleAddSeries = useCallback(() => {
-    navigation.navigate(ScreenType.SEARCH);
-  }, [navigation]);
+  const userInitials = userName
+    ? userName
+        .split(' ')
+        .map((word) => word[0])
+        .join('')
+        .toUpperCase()
+        .slice(0, 2)
+    : '?';
 
   useEffect(() => {
     if (!session?.user?.id) return;
 
     const fetchProfile = async () => {
-      setIsLoading(true);
       const { data, error } = await supabase
         .from('profiles')
         .select('first_name')
@@ -43,38 +35,16 @@ export const useViewModel = () => {
       if (!error && data) {
         setFirstName(data.first_name);
       }
-      setIsLoading(false);
     };
 
     fetchProfile();
   }, [session?.user?.id]);
 
   const greetingKey = getGreetingKey();
-  const filteredSeries = userSeries.filter((s) => s.status === activeStatus);
-
-  const statusCountMap: Record<SeriesStatus, number> = {
-    [SeriesStatus.Watching]: userSeries.filter(
-      (s) => s.status === SeriesStatus.Watching,
-    ).length,
-    [SeriesStatus.Completed]: userSeries.filter(
-      (s) => s.status === SeriesStatus.Completed,
-    ).length,
-    [SeriesStatus.Planned]: userSeries.filter(
-      (s) => s.status === SeriesStatus.Planned,
-    ).length,
-    [SeriesStatus.Dropped]: userSeries.filter(
-      (s) => s.status === SeriesStatus.Dropped,
-    ).length,
-  };
 
   return {
     firstName,
     greetingKey,
-    activeStatus,
-    setActiveStatus,
-    filteredSeries,
-    statusCountMap,
-    handleAddSeries,
-    isLoading,
+    userInitials,
   };
 };
