+++ src/views/profile/viewmodel.ts
@@ -1,11 +1,10 @@
 import { useAuth } from '@/context/AuthContext';
-import { useSeries } from '@/context/SeriesContext';
 import { useTranslation } from 'react-i18next';
 
 export const useViewModel = () => {
   const { t } = useTranslation();
   const { signOut, userName, userEmail } = useAuth();
-  const { userSeries } = useSeries();
+
 
   const userInitials = userName
     ? userName
